+++ deepparse/parser/address_parser.py
@@ -9,7 +9,10 @@ from torch.optim import SGD
 from torch.utils.data import DataLoader, Subset
 
 from . import formatted_parsed_address
+from .capturing import Capturing
 from .formatted_parsed_address import FormattedParsedAddress
+from .tools import validate_if_new_seq2seq_params, validate_if_new_prediction_tags, load_tuple_to_device, \
+    pretrained_parser_in_directory, get_files_in_directory, get_address_parser_in_directory, indices_splitting
 from ..converter import TagsConverter
 from ..converter import fasttext_data_padding, bpemb_data_padding, DataTransform
 from ..dataset_container import DatasetContainer
@@ -21,11 +24,11 @@ from ..fasttext_tools import download_fasttext_magnitude_embeddings
 from ..metrics import nll_loss, accuracy
 from ..network.bpemb_seq2seq import BPEmbSeq2SeqModel
 from ..network.fasttext_seq2seq import FastTextSeq2SeqModel
-from ..tools import CACHE_PATH, indices_splitting, load_tuple_to_device
+from ..preprocessing import AddressCleaner
+from ..tools import CACHE_PATH
 from ..vectorizer import FastTextVectorizer, BPEmbVectorizer
 from ..vectorizer import TrainVectorizer
 from ..vectorizer.magnitude_vectorizer import MagnitudeVectorizer
-from ..preprocessing import AddressCleaner
 
 _pre_trained_tags_to_idx = {
     "StreetNumber": 0,
@@ -173,9 +176,9 @@ class AddressParser:
 
         if path_to_retrained_model is not None:
             checkpoint_weights = torch.load(path_to_retrained_model, map_location="cpu")
-            if _validate_if_new_seq2seq_params(checkpoint_weights):
+            if validate_if_new_seq2seq_params(checkpoint_weights):
                 seq2seq_kwargs = checkpoint_weights.get("seq2seq_params")
-            if _validate_if_new_prediction_tags(checkpoint_weights):
+            if validate_if_new_prediction_tags(checkpoint_weights):
                 # We load the new tags_to_idx
                 tags_to_idx = checkpoint_weights.get("prediction_tags")
                 # We change the FIELDS for the FormattedParsedAddress
@@ -283,8 +286,7 @@ class AddressParser:
                 logging_path: str = "./checkpoints",
                 prediction_tags: Union[Dict, None] = None,
                 seq2seq_params: Union[Dict, None] = None) -> List[Dict]:
-
-        # pylint: disable=too-many-arguments, line-too-long, too-many-locals
+        # pylint: disable=too-many-arguments, line-too-long, too-many-locals, too-many-branches
         """
         Method to retrain the address parser model using a dataset with the same tags. We train using
         `experiment <https://poutyne.org/experiment.html>`_ from `poutyne <https://poutyne.org/index.html>`_
@@ -309,7 +311,10 @@ class AddressParser:
                 See Poutyne `callback <https://poutyne.org/callbacks.html#callback-class>`_ for more information. By
                 default, we set no callback.
             seed (int): Seed to use (by default 42).
-            logging_path (str): The logging path for the checkpoints. By default, the path is ``./checkpoints``.
+            logging_path (str): The logging path for the checkpoints. Poutyne will use the best one and reload the
+                state if any checkpoints are there. Thus, an error will be raised if you change the model type.
+                For example,  you retrain a FastText model and then retrain a BPEmb in the same logging path directory.
+                By default, the path is ``./checkpoints``.
             prediction_tags (Union[dict, None]): A dictionary where the keys are the address components
                 (e.g. street name) and the values are the components indices (from 0 to N + 1) to use during retraining
                 of a model. The ``+ 1`` corresponds to the End Of Sequence (EOS) token that needs to be included in the
@@ -338,6 +343,11 @@ class AddressParser:
             Due to pymagnitude, we could not train using the Magnitude embeddings, meaning it's not possible to
             train using the fasttext-light model. But, since we don't update the embeddings weights, one can retrain
             using the fasttext model and later on use the weights with the fasttext-light.
+
+        Note:
+            When retraining a model, Poutyne will create checkpoints. After the training, we use the best checkpoint
+            in a directory as the model to load. Thus, if you train two different models in the same directory,
+            the second retrain will not work due to model differences.
 
         Examples:
 
@@ -407,7 +417,7 @@ class AddressParser:
                     prediction_tags=address_components)
 
         """
-        if self.model_type == "fasttext-light":
+        if "fasttext-light" in self.model_type:
             raise ValueError("It's not possible to retrain a fasttext-light due to pymagnitude problem.")
 
         model_factory_dict = {"prediction_layer_len": 9}  # We set the default output dim size
@@ -456,25 +466,47 @@ class AddressParser:
                          loss_function=nll_loss,
                          batch_metrics=[accuracy])
 
-        train_res = exp.train(train_generator,
-                              valid_generator=valid_generator,
-                              epochs=epochs,
-                              seed=seed,
-                              callbacks=callbacks,
-                              verbose=self.verbose,
-                              disable_tensorboard=True)  # to remove tensorboard automatic logging
+        try:
+            # We capture poutyne print since it print some exception
+            with Capturing():
+                train_res = exp.train(train_generator,
+                                      valid_generator=valid_generator,
+                                      epochs=epochs,
+                                      seed=seed,
+                                      callbacks=callbacks,
+                                      verbose=self.verbose,
+                                      disable_tensorboard=True)  # to remove tensorboard automatic logging
+        except RuntimeError as error:
+            list_of_file_path = os.listdir(path='.')
+            if len(list_of_file_path) > 0:
+                if pretrained_parser_in_directory(logging_path):
+                    # Mean we might already have checkpoint in the training directory
+                    files_in_directory = get_files_in_directory(logging_path)
+                    retrained_address_parser_in_directory = get_address_parser_in_directory(
+                        files_in_directory)[0].split("_")[1]
+                    if self.model_type != retrained_address_parser_in_directory:
+                        raise ValueError(f"You are currently training a {self.model_type} in the directory "
+                                         f"{logging_path} where a different retrained "
+                                         f"{retrained_address_parser_in_directory} is currently his."
+                                         f" Thus, the loading of the model is failing. Change directory to retrain the"
+                                         f" {self.model_type}.") from error
+                    if self.model_type == retrained_address_parser_in_directory:
+                        raise ValueError(f"You are currently training a different {self.model_type} version from"
+                                         f" the one in the {logging_path}. Verify version.") from error
+            else:
+                raise RuntimeError(error) from error
+        else:
+            file_path = os.path.join(logging_path, f"retrained_{self.model_type}_address_parser.ckpt")
+            torch_save = {"address_tagger_model": exp.model.network.state_dict(), "model_type": self.model_type}
+            if seq2seq_params is not None:
+                # Means we have changed the seq2seq params
+                torch_save.update({"seq2seq_params": seq2seq_params})
+            if prediction_tags is not None:
+                #  Means we have changed the predictions tags
+                torch_save.update({"prediction_tags": prediction_tags})
 
-        file_path = os.path.join(logging_path, f"retrained_{self.model_type}_address_parser.ckpt")
-        torch_save = {"address_tagger_model": exp.model.network.state_dict(), "model_type": self.model_type}
-        if seq2seq_params is not None:
-            # Means we have changed the seq2seq params
-            torch_save.update({"seq2seq_params": seq2seq_params})
-        if prediction_tags is not None:
-            #  Means we have changed the predictions tags
-            torch_save.update({"prediction_tags": prediction_tags})
-
-        torch.save(torch_save, file_path)
-        return train_res
+            torch.save(torch_save, file_path)
+            return train_res
 
     def test(self,
              test_dataset_container: DatasetContainer,
@@ -540,7 +572,7 @@ class AddressParser:
                 address_parser.test(test_container) # Test the retrained model
 
         """
-        if self.model_type == "fasttext-light":
+        if "fasttext-light" in self.model_type:
             raise ValueError("It's not possible to test a fasttext-light due to pymagnitude problem. See Retrain method"
                              "doc for more details.")
 
@@ -727,11 +759,3 @@ class AddressParser:
         `"FastText"`.
         """
         return self._model_type_formatted
-
-
-def _validate_if_new_prediction_tags(checkpoint_weights: dict) -> bool:
-    return checkpoint_weights.get("prediction_tags") is not None
-
-
-def _validate_if_new_seq2seq_params(checkpoint_weights: dict) -> bool:
-    return checkpoint_weights.get("seq2seq_params") is not None