+++ deepparse/parser/capturing.py
@@ -1,0 +1,18 @@
+import sys
+from io import StringIO
+
+
+class Capturing(list):
+    """
+    Decorator function to capture the output of a function.
+    """
+
+    def __enter__(self):
+        self._stdout = sys.stdout
+        sys.stdout = self._stringio = StringIO()
+        return self
+
+    def __exit__(self, *args):
+        self.extend(self._stringio.getvalue().splitlines())
+        del self._stringio
+        sys.stdout = self._stdout