+++ tests/test_tools.py
@@ -8,8 +8,8 @@ from unittest.mock import patch
 
 import requests
 
-from deepparse import download_from_url, latest_version, download_weights, indices_splitting, \
-    handle_pre_trained_checkpoint, handle_poutyne_version, valid_poutyne_version
+from deepparse import download_from_url, latest_version, download_weights, handle_pre_trained_checkpoint, \
+    handle_poutyne_version, valid_poutyne_version
 from deepparse import handle_model_path, CACHE_PATH
 from tests.base_capture_output import CaptureOutputTestCase
 from tests.tools import create_file
@@ -168,27 +168,6 @@ class ToolsTests(CaptureOutputTestCase):
         with self.assertRaises(ValueError):
             bad_pickle_extension_checkpoint = "/a/path/to/a/model.pck"
             handle_model_path(bad_pickle_extension_checkpoint)
-
-    # test if splitting respect ratio splitting
-    def test_givenADataset_whenIndicesSplittingRatio8020_thenSplitIndices80Train20Valid(self):
-        number_of_data_points_in_dataset = 100
-        train_ratio = 0.8
-        expected_train_indices = [
-            83, 53, 70, 45, 44, 39, 22, 80, 10, 0, 18, 30, 73, 33, 90, 4, 76, 77, 12, 31, 55, 88, 26, 42, 69, 15, 40,
-            96, 9, 72, 11, 47, 85, 28, 93, 5, 66, 65, 35, 16, 49, 34, 7, 95, 27, 19, 81, 25, 62, 13, 24, 3, 17, 38, 8,
-            78, 6, 64, 36, 89, 56, 99, 54, 43, 50, 67, 46, 68, 61, 97, 79, 41, 58, 48, 98, 57, 75, 32, 94, 59
-        ]
-        expected_valid_indices = [63, 84, 37, 29, 1, 52, 21, 2, 23, 87, 91, 74, 86, 82, 20, 60, 71, 14, 92, 51]
-        expected_len_train_indices = 80
-        expected_len_valid_indices = 20
-
-        actual_train_indices, actual_valid_indices = indices_splitting(number_of_data_points_in_dataset,
-                                                                       train_ratio,
-                                                                       seed=self.a_seed)
-        self.assertEqual(len(actual_train_indices), expected_len_train_indices)
-        self.assertEqual(len(actual_valid_indices), expected_len_valid_indices)
-        self.assertEqual(actual_train_indices, expected_train_indices)
-        self.assertEqual(actual_valid_indices, expected_valid_indices)
 
     @patch("deepparse.tools.poutyne")
     def test_givenPoutyneVersionLowerThan12_givenHandlePreTrainedCheckpoint_thenRaiseError(self, poutyne_mock):